==> 04_Figure4/01_Fig_4_scrublet_prep.ipynb <==
import os
import shutil

# Get a list of directories
def directory_maker(base_dir):
    directory = [d for d in os.listdir(base_dir) if os.path.isdir(os.path.join(base_dir, d))]

    directory = [i for i in directory if i[0] != '.']
    return directory


destination_dir = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w1/"

base_directory1 = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/Exp00918_w1/outs/per_sample_outs/"

# Create destination directory if it doesn't exist
os.makedirs(destination_dir, exist_ok=True)

for directory in directory_maker(base_directory1):
    # Source file path
    source_file = os.path.join(base_directory1, directory,'count/', "sample_filtered_feature_bc_matrix.h5")

    # Destination file path with the name of the original directory
    destination_file = os.path.join(destination_dir, f"{directory}_sample_filtered_feature_bc_matrix.h5")

    # Copy file
    shutil.copy(source_file, destination_file)
    print(f"File copied from\n{source_file}\nto\n{destination_file}\n")

destination_dir = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w2/"

base_directory1 = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/Exp00918_w2/outs/per_sample_outs/"

# Create destination directory if it doesn't exist
os.makedirs(destination_dir, exist_ok=True)

for directory in directory_maker(base_directory1):
    # Source file path
    source_file = os.path.join(base_directory1, directory,'count/', "sample_filtered_feature_bc_matrix.h5")

    # Destination file path with the name of the original directory
    destination_file = os.path.join(destination_dir, f"{directory}_sample_filtered_feature_bc_matrix.h5")

    # Copy file
    shutil.copy(source_file, destination_file)
    print(f"File copied from\n{source_file}\nto\n{destination_file}\n")

destination_dir = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w3/"

base_directory1 = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/Exp00918_w3/outs/per_sample_outs/"

# Create destination directory if it doesn't exist
os.makedirs(destination_dir, exist_ok=True)

for directory in directory_maker(base_directory1):
    # Source file path
    source_file = os.path.join(base_directory1, directory,'count/', "sample_filtered_feature_bc_matrix.h5")

    # Destination file path with the name of the original directory
    destination_file = os.path.join(destination_dir, f"{directory}_sample_filtered_feature_bc_matrix.h5")

    # Copy file
    shutil.copy(source_file, destination_file)
    print(f"File copied from\n{source_file}\nto\n{destination_file}\n")

destination_dir = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w4/"

base_directory1 = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/Exp00918_w4/outs/per_sample_outs/"

# Create destination directory if it doesn't exist
os.makedirs(destination_dir, exist_ok=True)

for directory in directory_maker(base_directory1):
    # Source file path
    source_file = os.path.join(base_directory1, directory,'count/', "sample_filtered_feature_bc_matrix.h5")

    # Destination file path with the name of the original directory
    destination_file = os.path.join(destination_dir, f"{directory}_sample_filtered_feature_bc_matrix.h5")

    # Copy file
    shutil.copy(source_file, destination_file)
    print(f"File copied from\n{source_file}\nto\n{destination_file}\n")
