==> 04_Figure4/02_Fig_4_scrublet.ipynb <==
!pip install scrublet

import sys
import getopt
import shutil
import h5py
import pandas
import scanpy as sc
import scrublet
import re
import os

path = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w1"
paths = [path+i for i in os.listdir(path) if i[0] != '.']
paths

def scrubby(x):
    name = x.split('/')[-1].split('_sample')[0]
    
    if x.split('.')[1] == 'h5':
        adata = sc.read_10x_h5(x)
        
        scrub = scrublet.Scrublet(adata.X)
        doublet_scores, predicted_doublets = scrub.scrub_doublets(verbose=False)

        df = adata.obs
        df['Doublet_Score'] = doublet_scores
        df['Predicted_Doublet'] = predicted_doublets

        df.to_csv(path+name+'.csv')


path = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w1/"
paths = [path+i for i in os.listdir(path) if i[0] != '.']

for i in paths: scrubby(i)

path = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w2/"
paths = [path+i for i in os.listdir(path) if i[0] != '.']

for i in paths: scrubby(i)

path = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w3/"
paths = [path+i for i in os.listdir(path) if i[0] != '.']

for i in paths: scrubby(i)

path = "/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w4/"
paths = [path+i for i in os.listdir(path) if i[0] != '.']

for i in paths: scrubby(i)
