==> 04_Figure4/03_Fig_4_n64_preprocessing.ipynb <==
# Exp00918 n=64 Validation Run

library(dplyr)
library(Seurat)
library(patchwork)
library(H5weaver)
library(hise)
library(tidyverse)
library(SeuratObject)
library(ggrepel)

h5s_w1 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w1',
    pattern = 'filtered_feature_bc_matrix.h5$', 
    full.names = TRUE, 
    recursive = TRUE
)

scrub_w1 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w1',
    pattern = '.csv$', 
    full.names = TRUE, 
    recursive = TRUE
) %>% 
    lapply(read.csv) %>% 
    lapply(function(x){
        rownames(x) <- x$X
        x$X <- NULL
        x$Barcodes <- rownames(x)
        return(x)
    })


h5s_w2 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w2',
    pattern = 'filtered_feature_bc_matrix.h5$',
    full.names = TRUE, 
    recursive = TRUE
)

scrub_w2 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w2',
    pattern = '.csv$',
    full.names = TRUE, 
    recursive = TRUE
) %>% 
    lapply(read.csv) %>% 
    lapply(function(x){
        rownames(x) <- x$X
        x$X <- NULL
        x$Barcodes <- rownames(x)
        return(x)
    })


h5s_w3 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w3',
    pattern = 'filtered_feature_bc_matrix.h5$', 
    full.names = TRUE, 
    recursive = TRUE
)

scrub_w3 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w3',
    pattern = '.csv$',
    full.names = TRUE, 
    recursive = TRUE
) %>% 
    lapply(read.csv) %>% 
    lapply(function(x){
        rownames(x) <- x$X
        x$X <- NULL
        x$Barcodes <- rownames(x)
        return(x)
    })


h5s_w4 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w4',
    pattern = 'filtered_feature_bc_matrix.h5$', 
    full.names = TRUE, 
    recursive = TRUE
)

scrub_w4 <- list.files(
    path = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/scrublet_fig4/w4',
    pattern = '.csv$', 
    full.names = TRUE, 
    recursive = TRUE
) %>% 
    lapply(read.csv) %>% 
    lapply(function(x){
        rownames(x) <- x$X
        x$X <- NULL
        x$Barcodes <- rownames(x)
        return(x)
    })


so_gen <- function(x){
    # pro <- strsplit(strsplit(x,'/per_sample_outs/')[[1]][2],'/count/')[[1]][1]
    pro <- strsplit(x, '/')[[1]][10] 
    pro <- strsplit(pro, '_')[[1]][1]
    
    exp <- strsplit(x, '/')[[1]][9] 

    pro <- paste(pro,exp,sep='_')

    mtx <- Read10X_h5(x) 
    so <- CreateSeuratObject(mtx,project=pro) 
    return(so) 
    }

fixed_so_w1 <- lapply(h5s_w1, so_gen)
fixed_so_w2 <- lapply(h5s_w2, so_gen)
fixed_so_w3 <- lapply(h5s_w3, so_gen)
fixed_so_w4 <- lapply(h5s_w4, so_gen)


scrub.it <- function(so_l, scrub_l){
    for (i in c(1:length(so_l))){
        so_l[[i]][[]]$Barcodes <- rownames(so_l[[i]][[]])
        # so_l[[i]][[]]$Barcodes <- rownames(scrub_l[[i]])
        so_l[[i]][[]] <- left_join(so_l[[i]][[]], scrub_l[[i]], by = 'Barcodes') # , by='row.names', all=TRUE
    }
    return(so_l)
}

fixed_so_w1 <- fixed_so_w1 %>% scrub.it(scrub_w1)
fixed_so_w2 <- fixed_so_w2 %>% scrub.it(scrub_w2)
fixed_so_w3 <- fixed_so_w3 %>% scrub.it(scrub_w3)
fixed_so_w4 <- fixed_so_w4 %>% scrub.it(scrub_w4)

fully_w1 <- Reduce(merge,fixed_so_w1) %>% JoinLayers()
fully_w2 <- Reduce(merge,fixed_so_w2) %>% JoinLayers()
fully_w3 <- Reduce(merge,fixed_so_w3) %>% JoinLayers()
fully_w4 <- Reduce(merge,fixed_so_w4) %>% JoinLayers()


rm(fixed_so_w1)
rm(fixed_so_w2)
rm(fixed_so_w3)
rm(fixed_so_w4)

fully_w1[["percent.mt"]] <- PercentageFeatureSet(fully_w1, pattern = "^MT-")
fully_w2[["percent.mt"]] <- PercentageFeatureSet(fully_w2, pattern = "^MT-")
fully_w3[["percent.mt"]] <- PercentageFeatureSet(fully_w3, pattern = "^MT-")
fully_w4[["percent.mt"]] <- PercentageFeatureSet(fully_w4, pattern = "^MT-")


fully_w1 <- subset(fully_w1, subset = percent.mt < 5)
fully_w2 <- subset(fully_w2, subset = percent.mt < 5)
fully_w3 <- subset(fully_w3, subset = percent.mt < 5)
fully_w4 <- subset(fully_w4, subset = percent.mt < 5)


fully_w1 <- fully_w1 %>% subset(subset=Predicted_Doublet=='False')
fully_w2 <- fully_w2 %>% subset(subset=Predicted_Doublet=='False')
fully_w3 <- fully_w3 %>% subset(subset=Predicted_Doublet=='False')
fully_w4 <- fully_w4 %>% subset(subset=Predicted_Doublet=='False')


## Plotting

fully_w1$well <- substr(fully_w1$orig.ident, 13,14)
fully_w2$well <- substr(fully_w2$orig.ident, 13,14)
fully_w3$well <- substr(fully_w3$orig.ident, 13,14)
fully_w4$well <- substr(fully_w4$orig.ident, 13,14)

fully_w1$donor <- substr(fully_w1$orig.ident,1,7)
fully_w2$donor <- substr(fully_w2$orig.ident,1,7)
fully_w3$donor <- substr(fully_w3$orig.ident,1,7)
fully_w4$donor <- substr(fully_w4$orig.ident,1,7)

fully_w1$donor2 <- substr(fully_w1$orig.ident,1,11)
fully_w2$donor2 <- substr(fully_w2$orig.ident,1,11)
fully_w3$donor2 <- substr(fully_w3$orig.ident,1,11)
fully_w4$donor2 <- substr(fully_w4$orig.ident,1,11)


head(fully_w1)

fully <- merge(fully_w1, c(fully_w2, fully_w3, fully_w4)) %>% JoinLayers()


head(fully)

fully <- NormalizeData(fully) %>% 
    FindVariableFeatures() %>% 
    ScaleData() %>% 
    RunPCA() %>% 
    RunUMAP(dims = 1:20) %>% 
    FindNeighbors(dims = 1:20) %>% 
    FindClusters(resolution = 0.5)

saveRDS(fully, '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/Fig_4_Final.rds')

head(fully)

sessionInfo()
