==> 04_Figure4/05_palmo_analysis_clean.ipynb <==
# Libraries

# sudo apt install libcurl4-openssl-dev 

# install.packages('curl')

# install.packages("https://cran.r-project.org/src/contrib/Archive/MASS/MASS_7.3-60.tar.gz", repos=NULL)

# install.packages("https://cran.r-project.org/src/contrib/Archive/Matrix/Matrix_1.6-5.tar.gz", repos=NULL)

# install.packages("https://cran.r-project.org/src/contrib/Archive/curl/curl_4.3.tar.gz", repos=NULL)

install.packages('Seurat')

# install.packages("devtools")

# devtools::install_github("r-lib/rlang")

# install.packages('PALMO')
library("devtools")
install_github("aifimmunology/PALMO")

# Palmo

library(Seurat)
library(dplyr)
library(PALMO)

so <- readRDS(file = '/home/jupyter/CS15_WHBL/CWB_Paper/01_Final_Data/04_Data/Fig_4_Final.rds')

mtx <- so[['RNA']]$counts

meta <- so@meta.data

# mtx <- readRDS('exp918_mtx.rds')
# meta <- readRDS('exp918_meta.rds')

so <- CreateSeuratObject(counts = mtx, meta.data = meta)

so

head(so)

so <- NormalizeData(so) %>% FindVariableFeatures() %>% ScaleData() %>% RunPCA()

so_split <- SplitObject(so, split.by='donor')

ct_label_transfer <- function(so){
    ref <- SeuratDisk::LoadH5Seurat(file = '/home/jupyter/pbmc_multimodal.h5seurat')
    
    anchors <- FindTransferAnchors(
        reference = ref,
        query = so,
        normalization.method = "SCT",
        reference.reduction = "spca",
        dims = 1:50)
    
    so <- TransferData(anchorset = anchors, reference = ref, query = so,
          refdata = list(
          celltype.l1 = "celltype.l1",
          celltype.l2 = "celltype.l2",
          celltype.l3 = "celltype.l3")
        )
    
    return(so)

}

# ref <- SeuratDisk::LoadH5Seurat(file = '/home/jupyter/pbmc_multimodal.h5seurat')

# anchors <- FindTransferAnchors(
#     reference = ref,
#     query = so,
#     normalization.method = "SCT",
#     reference.reduction = "spca",
#     dims = 1:50)

# so <- TransferData(anchorset = anchors, reference = ref, query = so,
#               refdata = list(
#               celltype.l1 = "celltype.l1",
#               celltype.l2 = "celltype.l2",
#               celltype.l3 = "celltype.l3"))

rm(so)

# so <- ct_label_transfer(so)
so_split <- so_split %>% lapply(ct_label_transfer)


so_split

so_split1 <- so_split[[1]]
so_split2 <- so_split[[2]]
so_split3 <- so_split[[3]]
so_split4 <- so_split[[4]]
so_split5 <- so_split[[5]]
# so_split6 <- so_split[[6]]
# so_split7 <- so_split[[7]]
# so_split8 <- so_split[[8]]

rm(so_split)

so_split12 <- merge(so_split1,so_split2)
rm(so_split1)
rm(so_split2)

so_split34 <- merge(so_split3,so_split4)
rm(so_split3)
rm(so_split4)

so_split1234 <- merge(so_split12, so_split34)
rm(so_split12)
rm(so_split34)

so <- merge(so_split1234, so_split5)
rm(so_split5)
rm(so_split1234)

# so <- Reduce(merge, so_split)
# rm(so_split)

head(so@meta.data)

table(so$donor)

table(so$donor2)

so$Sample <- so$donor2
so$Replicate <- so$donor2
so$Pool_Well <- so$well
so$Donor <- so$donor

palmo_obj <- createPALMOobject(anndata=so@meta.data, data=so)

palmo_obj <- annotateMetadata(data_object=palmo_obj,
    sample_column= "Sample",
    donor_column= "Donor",
    time_column= "Pool_Well")

palmo_obj <- mergePALMOdata(data_object=palmo_obj, datatype="singlecell")

palmo_obj <- avgExpCalc(data_object=palmo_obj,
assay="RNA", group_column="predicted.celltype.l2")
head(palmo_obj@curated[["anndata"]]) #merged annotation data
head(palmo_obj@curated[["data"]]) #scRNA average expression data

# options(repr.plot.width = 20, repr.plot.height = 8)
# palmo_obj <- cvCalcSCProfile(data_object=palmo_obj,
#     housekeeping_genes= "ACTB",
#     fileName="scrna")

options(repr.plot.width = 20, repr.plot.height = 8)
palmo_obj <- cvCalcSCProfile(data_object=palmo_obj, meanThreshold = 0.1,
    housekeeping_genes= "ACTB",
    fileName="scrna")

#Check the group of interest
head(palmo_obj@curated$anndata)

#Variance decomposition
featureSet <- c("Replicate", "Donor", "Pool_Well")

palmo_obj <- lmeVariance(data_object=palmo_obj,
    featureSet=featureSet,
    meanThreshold=0.05, cl=4,
    fileName="scrna")

#Get the result
var_decomp <- palmo_obj@result$variance_decomposition
var_decomp

saveRDS(so, '/home/jupyter/CS15_WHBL/CWB_Paper/Palmo_so.rds')
